bg-remover-app/backend/requirements.txt
Flask
flask-cors
rembg
Pillow
